Apply dtype change
i saw it in the parselmouth thing might as well put it in to make sure

+++ modules/pe/pw.py
@@ -12,7 +12,7 @@ class HarvestPE(BasePE):
         f0_floor = hparams['fmin']
         f0_ceil = hparams['fmax']
 
-        f0, _ = pw.harvest(waveform, hparams['audio_sample_rate'], f0_floor=f0_floor, f0_ceil=f0_ceil, frame_period=time_step)
+        f0, _ = pw.harvest(waveform, hparams['audio_sample_rate'], f0_floor=f0_floor, f0_ceil=f0_ceil, frame_period=time_step).astype(np.float32)
         f0 = pad_frames(f0, hop_size, waveform.shape[0], length)
         uv = f0 == 0
 
@@ -29,7 +29,7 @@ class DioPE(BasePE):
         f0_ceil = hparams['fmax']
 
         f0, t = pw.dio(waveform, hparams['audio_sample_rate'], f0_floor=f0_floor, f0_ceil=f0_ceil, frame_period=time_step)
-        f0 = pw.stonemask(waveform, f0, t, hparams['audio_sample_rate'])
+        f0 = pw.stonemask(waveform, f0, t, hparams['audio_sample_rate']).astype(np.float32)
         f0 = pad_frames(f0, hop_size, waveform.shape[0], length)
         uv = f0 == 0
